fix: extract b0 from dwi.h5
